+++ src/collective/recipe/backup/tests/test_docs.py
@@ -13,6 +13,7 @@ import collective.recipe.backup
 import doctest
 import os
 import re
+import sys
 import tempfile
 import unittest
 import zc.buildout.testing
@@ -101,8 +102,14 @@ def test_suite():
         'options.rst',
         'prefix.rst',
         'zipbackup.rst',
-        'zope2instance.rst',
     ]
+    test_file = 'zope2instance.rst'
+    if sys.version_info[0] > 2:
+        print('WARNING: ignoring {0} tests on Python 3.'.format(test_file))
+        print('Enable them when there is a compatible mailinglogger version.')
+        print('See https://github.com/collective/collective.recipe.backup/issues/31')
+    else:
+        docfiles.append(test_file)
     for docfile in docfiles:
         suite.addTest(doctest.DocFileSuite(
             docfile,